requirements.txt
torch
torchvision
numpy
matplotlib
pandas
scikit-learn
Pillow
tensorflow
tensorflow_hub
diffusers
transformers
accelerate
safetensors
librosa
soundfile
scikit-image
